pyproject.toml
[project]
name = "krakem-main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "asyncio>=4.0.0",
    "django>=6.0.2",
    "django-extensions>=4.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pyjwt>=2.11.0",
    "pytest>=9.0.2",
]
